+++ setup.cfg
@@ -84,6 +84,7 @@ exclude =
     thinc/__init__.py
     thinc/layers/__init__.py
     thinc/shims/__init__.py
+    thinc/api.py
 
 [tool:pytest]
 markers =
