+++ thinc/__init__.py
@@ -2,6 +2,3 @@
 import numpy
 
 from .about import __version__
-from .config import Config
-from .util import prefer_gpu, require_gpu
-from ._registry import registry
