+++ thinc/backends/__init__.py
@@ -4,6 +4,7 @@ from .cupy_ops import CupyOps  # noqa: F401
 from .numpy_ops import NumpyOps  # noqa: F401
 from ..util import create_thread_local, get_ops
 
+
 STATE = create_thread_local({"Ops": NumpyOps, "ops": NumpyOps()})
