+++ thinc/config.py
@@ -1,8 +1,14 @@
-from typing import Union, Dict, Any, Optional, List, Tuple
+from typing import Union, Dict, Any, Optional, List, Tuple, Callable, Type
+from types import GeneratorType
 from configparser import ConfigParser, ExtendedInterpolation
-import io
 from pathlib import Path
+from pydantic import BaseModel, create_model, ValidationError
+from pydantic.main import ModelMetaclass
+from wasabi import table
 import srsly
+import catalogue
+import inspect
+import io
 
 
 class Config(dict):
@@ -85,4 +91,223 @@ class Config(dict):
         return self.from_str(text)
 
 
-__all__ = ["Config"]
+class ConfigValidationError(ValueError):
+    def __init__(
+        self,
+        config: Union[Config, Dict[str, Dict[str, Any]]],
+        errors: List[Dict[str, Any]],
+        message: str = "Config validation error",
+    ) -> None:
+        """Custom error for validating configs."""
+        data = []
+        for error in errors:
+            err_loc = " -> ".join([str(p) for p in error.get("loc", [])])
+            data.append((err_loc, error.get("msg")))
+        result = [message, table(data), f"{config}"]
+        ValueError.__init__(self, "\n\n" + "\n".join(result))
+
+
+class EmptySchema(BaseModel):
+    class Config:
+        extra = "allow"
+        arbitrary_types_allowed = True
+
+
+class _PromiseSchemaConfig:
+    extra = "forbid"
+    arbitrary_types_allowed = True
+
+
+class registry(object):
+    optimizers = catalogue.create("thinc", "optimizers", entry_points=True)
+    schedules = catalogue.create("thinc", "schedules", entry_points=True)
+    layers = catalogue.create("thinc", "layers", entry_points=True)
+
+    @classmethod
+    def create(cls, registry_name: str, entry_points: bool = False) -> None:
+        """Create a new custom registry."""
+        if hasattr(cls, registry_name):
+            raise ValueError(f"Registry '{registry_name}' already exists")
+        reg = catalogue.create("thinc", registry_name, entry_points=entry_points)
+        setattr(cls, registry_name, reg)
+
+    @classmethod
+    def get(cls, registry_name: str, func_name: str) -> Callable:
+        """Get a registered function from a given registry."""
+        if not hasattr(cls, registry_name):
+            raise ValueError(f"Unknown registry: '{registry_name}'")
+        reg = getattr(cls, registry_name)
+        func = reg.get(func_name)
+        if func is None:
+            raise ValueError(f"Could not find '{func_name}' in '{registry_name}'")
+        return func
+
+    @classmethod
+    def make_from_config(
+        cls,
+        config: Union[Config, Dict[str, Dict[str, Any]]],
+        *,
+        schema: Type[BaseModel] = EmptySchema,
+        validate: bool = True,
+    ) -> Config:
+        """Unpack a config dictionary, creating objects from the registry
+        recursively. If validate=True, the config will be validated against the
+        type annotations of the registered functions referenced in the config
+        (if available) and/or the schema (if available).
+        """
+        # Valid: {"optimizer": {"@optimizers": "my_cool_optimizer", "rate": 1.0}}
+        # Invalid: {"@optimizers": "my_cool_optimizer", "rate": 1.0}
+        if cls.is_promise(config):
+            err_msg = "The top-level config object can't be a reference to a registered function."
+            raise ConfigValidationError(config, [{"msg": err_msg}])
+        _, resolved = cls._fill(config, schema, validate)
+        return resolved
+
+    @classmethod
+    def fill_config(
+        cls,
+        config: Union[Config, Dict[str, Dict[str, Any]]],
+        *,
+        schema: Type[BaseModel] = EmptySchema,
+        validate: bool = True,
+    ) -> Config:
+        """Unpack a config dictionary, leave all references to registry
+        functions intact and don't resolve them, but fill in all values and
+        defaults based on the type annotations. If validate=True, the config
+        will be validated against the type annotations of the registered
+        functions referenced in the config (if available) and/or the schema
+        (if available).
+        """
+        # Valid: {"optimizer": {"@optimizers": "my_cool_optimizer", "rate": 1.0}}
+        # Invalid: {"@optimizers": "my_cool_optimizer", "rate": 1.0}
+        if cls.is_promise(config):
+            err_msg = "The top-level config object can't be a reference to a registered function."
+            raise ConfigValidationError(config, [{"msg": err_msg}])
+        filled, _ = cls._fill(config, schema, validate)
+        return filled
+
+    @classmethod
+    def _fill(
+        cls,
+        config: Union[Config, Dict[str, Dict[str, Any]]],
+        schema: Type[BaseModel] = EmptySchema,
+        validate: bool = True,
+    ) -> Tuple[Config, Config]:
+        """Build two representations of the config: one where the promises are
+        preserved, and a second where the promises are represented by their
+        return types. Use the validation representation to get default
+        values via pydantic. The defaults are filled into both representations.
+        """
+        filled: Dict[str, Any] = {}
+        validation: Dict[str, Any] = {}
+        for key, value in config.items():
+            if cls.is_promise(value):
+                promise_schema = cls.make_promise_schema(value)
+                filled[key], _ = cls._fill(value, promise_schema, validate)
+                # Call the function and populate the field value. We can't just
+                # create an instance of the type here, since this wouldn't work
+                # for generics / more complex custom types
+                getter = cls.get_constructor(filled[key])
+                args, kwargs = cls.parse_args(filled[key])
+                try:
+                    validation[key] = getter(*args, **kwargs)
+                except TypeError as err:
+                    err_msg = "Can't construct config: calling registry function failed"
+                    raise ConfigValidationError(
+                        {key: value}, [{"msg": err, "loc": [getter.__name__]}], err_msg
+                    )
+                if isinstance(validation[key], GeneratorType):
+                    # If value is a generator we can't validate type without
+                    # consuming it (which doesn't work if it's infinite – see
+                    # schedule for examples). So we skip it.
+                    validation[key] = []
+            elif hasattr(value, "items"):
+                field_type = EmptySchema
+                if key in schema.__fields__:
+                    field = schema.__fields__[key]
+                    field_type = field.type_
+                    if not isinstance(field.type_, ModelMetaclass):
+                        # If we don't have a pydantic schema and just a type
+                        field_type = EmptySchema
+                filled[key], validation[key] = cls._fill(value, field_type, validate)
+            else:
+                filled[key] = value
+                validation[key] = value
+        # Now that we've filled in all of the promises, update with defaults
+        # from schema, and validate if validation is enabled
+        if validate:
+            try:
+                result = schema.parse_obj(validation)
+            except ValidationError as e:
+                raise ConfigValidationError(config, e.errors())
+        else:
+            # Same as parse_obj, but without validation
+            result = schema.construct(**validation)
+        validation.update(result.dict())
+        for key, value in validation.items():
+            if key not in filled:
+                filled[key] = value
+        return Config(filled), Config(validation)
+
+    @classmethod
+    def is_promise(cls, obj: Any) -> bool:
+        """Check whether an object is a "promise", i.e. contains a reference
+        to a registered function (via a key starting with `"@"`.
+        """
+        if not hasattr(obj, "keys"):
+            return False
+        id_keys = [k for k in obj.keys() if k.startswith("@")]
+        if len(id_keys) != 1:
+            return False
+        else:
+            return True
+
+    @classmethod
+    def get_constructor(cls, obj: Dict[str, Any]) -> Callable:
+        id_keys = [k for k in obj.keys() if k.startswith("@")]
+        if len(id_keys) != 1:
+            err = f"A block can only contain one function registry reference. Got: {id_keys}"
+            raise ValueError(err)
+        else:
+            key = id_keys[0]
+            value = obj[key]
+            return cls.get(key[1:], value)
+
+    @classmethod
+    def parse_args(cls, obj: Dict[str, Any]) -> Tuple[List[Any], Dict[str, Any]]:
+        args = []
+        kwargs = {}
+        for key, value in obj.items():
+            if not key.startswith("@"):
+                if isinstance(key, int) or key.isdigit():
+                    args.append((int(key), value))
+                else:
+                    kwargs[key] = value
+        return [value for key, value in sorted(args)], kwargs
+
+    @classmethod
+    def make_promise_schema(cls, obj: Dict[str, Any]) -> Type[BaseModel]:
+        """Create a schema for a promise dict (referencing a registry function)
+        by inspecting the function signature.
+        """
+        func = cls.get_constructor(obj)
+        # Read the argument annotations and defaults from the function signature
+        id_keys = [k for k in obj.keys() if k.startswith("@")]
+        sig_args: Dict[str, Any] = {id_keys[0]: (str, ...)}
+        for param in inspect.signature(func).parameters.values():
+            annotation = param.annotation if param.annotation != param.empty else Any
+            # If no default value is specified assume that it's required
+            if param.default != param.empty:
+                sig_args[param.name] = (annotation, param.default)
+            else:
+                sig_args[param.name] = (annotation, ...)
+        sig_args["__config__"] = _PromiseSchemaConfig
+        return create_model("ArgModel", **sig_args)
+
+    @classmethod
+    def get_return_type(cls, obj: Dict[str, Any]):
+        func = cls.get_constructor(obj)
+        return inspect.signature(func).return_annotation
+
+
+__all__ = ["Config", "registry"]
