+++ thinc/types.py
@@ -571,24 +571,3 @@ class Padded:
 
     data: Array
     size_at_t: Array
-
-
-__all__ = [
-    "Array",
-    "Floats1d",
-    "Floats2d",
-    "Floats3d",
-    "Floats4d",
-    "FloatsNd",
-    "Ints1d",
-    "Ints2d",
-    "Ints3d",
-    "Ints4d",
-    "IntsNd",
-    "RNNState",
-    "Ragged",
-    "Padded",
-    "Xp",
-    "Shape",
-    "DTypes",
-]
